+++ shopshoe_fe/src/context/Size.tsx
@@ -4,6 +4,7 @@ import { Size } from "../interface/Size";
 import sizeReducer from "../reducers/SizeReducer";
 import { toast } from "react-toastify";
 import { useNavigate } from "react-router-dom";
+import Swal from "sweetalert2";
 export type SizeContextType = {
     stateSize: { size: Size[] };
     dispatch: React.Dispatch<any>;
@@ -14,11 +15,13 @@ export const SizeContext = createContext({} as SizeContextType);
 export const SizeProvider = ({ children }: { children: React.ReactNode }) => {
     const [stateSize, dispatch] = useReducer(sizeReducer, { size: [] });
     const nav = useNavigate();
+
+    const getSize = async () => {
+        const { data } = await instance.get("/size");
+        dispatch({ type: "GET_SIZE", payload: data.data });
+    };
     useEffect(() => {
-        (async () => {
-            const { data } = await instance.get("/size");
-            dispatch({ type: "GET_SIZE", payload: data.data });
-        })();
+        getSize();
     }, []);
     const CreateSize = async (dataSize: Size) => {
         try {
@@ -32,13 +35,51 @@ export const SizeProvider = ({ children }: { children: React.ReactNode }) => {
         }
     };
     const DeleteSize = async (_id: Size | undefined) => {
+        // try {
+        //     await instance.delete(`/size/${_id}`);
+        //     dispatch({ type: "REMOVE_SIZE", payload: _id });
+        //     toast.success("Xóa size thành công");
+        //     window.location.reload();
+        // } catch (error) {
+        //     toast.error("Xóa size thất bại");
+        // }
+        const swalWithBootstrapButtons = Swal.mixin({
+            customClass: {
+                confirmButton: "custom-confirm-button",
+                cancelButton: "custom-cancel-button",
+            },
+            buttonsStyling: false,
+        });
+
         try {
-            await instance.delete(`/size/${_id}`);
-            dispatch({ type: "REMOVE_SIZE", payload: _id });
-            toast.success("Xóa size thành công");
-            window.location.reload();
+            const result = await swalWithBootstrapButtons.fire({
+                title: "Are you sure?",
+                text: "You won't be able to revert this!",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonText: "Yes, delete it!",
+                cancelButtonText: "No, cancel!",
+                reverseButtons: true,
+            });
+
+            if (result.isConfirmed) {
+                await instance.delete(`/size/${_id}`);
+                dispatch({ type: "REMOVE_SIZE", payload: _id });
+                getSize();
+                swalWithBootstrapButtons.fire({
+                    title: "Deleted!",
+                    text: "Your file has been deleted.",
+                    icon: "success",
+                });
+            } else if (result.dismiss === Swal.DismissReason.cancel) {
+                swalWithBootstrapButtons.fire({
+                    title: "Cancelled",
+                    text: "Your file is safe :)",
+                    icon: "error",
+                });
+            }
         } catch (error) {
-            toast.error("Xóa size thất bại");
+            toast.error("Không thể xóa");
         }
     };
     return (
